+++ backend/src/models/user_model.js
@@ -1,8 +1,8 @@
-const pool = require("../config/db");
+const pool = require("../config/database_config");
 
 const findByEmailOrUsername = async (email, username) => {
   const result = await pool.query(
-    "SELECT * FROM users WHERE email = $1 OR username = $2",
+    "SELECT * FROM auth.users WHERE email = $1 OR username = $2",
     [email, username]
   );
   return result.rows;
@@ -17,7 +17,7 @@ const createUser = async (data) => {
   const placeholders = keys.map((_, i) => `$${i + 1}`);
 
   const query = `
-    INSERT INTO users (${keys.join(", ")})
+    INSERT INTO auth.users (${keys.join(", ")})
     VALUES (${placeholders.join(", ")})
     RETURNING *;
   `;
@@ -27,7 +27,7 @@ const createUser = async (data) => {
 };
 
 const findByUsername = async (username) => {
-  const result = await pool.query("SELECT * FROM users WHERE username = $1", [
+  const result = await pool.query("SELECT * FROM auth.users WHERE username = $1", [
     username,
   ]);
   return result.rows[0];
@@ -35,7 +35,7 @@ const findByUsername = async (username) => {
 
 const getUsers = async () => {
   const result = await pool.query
-  ("SELECT u.user_id, u.username, u.email, u.phone, u.created_at, r.role_name, r.role_id FROM users u JOIN roles r ON u.role_id = r.role_id ORDER BY u.created_at DESC");
+  ("SELECT u.user_id, u.username, u.email, u.phone, u.created_at, r.role_name, r.role_id FROM auth.users u JOIN auth.roles r ON u.role_id = r.role_id ORDER BY u.created_at DESC");
   return result.rows;
 };
 
@@ -48,7 +48,7 @@ const updateUser = async (id, fields) => {
   values.push(id);
 
   const query = `
-    UPDATE users 
+    UPDATE auth.users 
     SET ${setQuery}
     WHERE user_id = $${values.length}
     RETURNING user_id, username, email, phone, role_id
@@ -57,21 +57,22 @@ const updateUser = async (id, fields) => {
   const result = await pool.query(query, values);
   return result.rows[0];
 };
+
 const deleteUser = async (id) => {
   const result = await pool.query(
-    "DELETE FROM users WHERE user_id = $1 RETURNING user_id, username, email",
+    "DELETE FROM auth.users WHERE user_id = $1 RETURNING user_id, username, email",
     [id]
   );
   return result.rows[0];
 };
 
-const updateUsername = async (id, username) => {
-  const result = await pool.query(
-    "UPDATE users SET username = $1 WHERE user_id = $2 RETURNING *",
-    [username, id]
-  );
-  return result.rows[0];
-};
+// const updateUsername = async (id, username) => {
+//   const result = await pool.query(
+//     "UPDATE auth.users SET username = $1 WHERE user_id = $2 RETURNING *",
+//     [username, id]
+//   );
+//   return result.rows[0];
+// };
 
 module.exports = {
   findByEmailOrUsername,
@@ -80,5 +81,5 @@ module.exports = {
   getUsers,
   updateUser,
   deleteUser,
-  updateUsername,
+  // updateUsername,
 };
